@@ -33,9 +33,11 @@
   "dependencies": {
     "dotenv": "^8.2.0",
     "html-webpack-plugin": "^4.5.1",
+    "mobile-device-detect": "^0.4.3",
     "node-sass": "^5.0.0",
     "tslint": "^6.1.3",
     "typed-js": "^0.2.3",
+    "webpack-modernizr-loader": "^5.0.0",
     "typed.js": "^2.0.11"
   },
   "devDependencies": {
@@ -79,7 +81,7 @@
     "ts-node": "^9.1.1",
     "tslint-config-prettier": "^1.18.0",
     "typescript": "^4.1.3",
-    "webpack": "^5.16.0",
+    "webpack": "^5.17.0",
     "webpack-cli": "^4.4.0",
     "webpack-dev-server": "^3.11.2"
   }
