@@ -1,8 +1,8 @@
-import { Entity} from "../utils";
-import {Player} from "../Player";
+import { Entity } from "../utils"
+import { Player } from "../Player"
 
 export class Bombs extends Entity {
-    constructor(public readonly Factory: Player) {
-        super()
-    }
+  constructor(public readonly Factory: Player) {
+    super()
+  }
 }
