@@ -1,9 +1,9 @@
-import { CanvasLayer } from "../canvas-layer";
-import { Canvas } from "../../utils";
+import { CanvasLayer } from "../canvas-layer"
+import { Canvas } from "../../utils"
 
-jest.mock('../../utils')
-describe('>>> CanvasLayer', () => {
-  it('should create Background canvas only once', () => {
+jest.mock("../../utils")
+describe(">>> CanvasLayer", () => {
+  it("should create Background canvas only once", () => {
     expect(Canvas).not.toBeCalled()
 
     const canvas1 = CanvasLayer.Background
@@ -13,7 +13,7 @@ describe('>>> CanvasLayer', () => {
     expect(Canvas).toBeCalledTimes(1)
   })
 
-  it('should create Foreground canvas only once', () => {
+  it("should create Foreground canvas only once", () => {
     expect(Canvas).not.toBeCalled()
 
     const canvas1 = CanvasLayer.Foreground
