@@ -1,8 +1,8 @@
-import { Grid } from '../grid'
-import { Node } from "../../node";
-import { Settings } from "../../settings";
+import { Grid } from "../grid"
+import { Node } from "../../node"
+import { Settings } from "../../settings"
 
-describe('Grid', () => {
+describe("Grid", () => {
   const nodeCount = Settings.grid.dimension * Settings.grid.dimension
   let grid: Grid
 
@@ -10,9 +10,9 @@ describe('Grid', () => {
     grid = new Grid()
   })
 
-  it('should awake and update all children', () => {
-    const spyNodeAwake = jest.spyOn(Node.prototype, 'Awake')
-    const spyNodeUpdate = jest.spyOn(Node.prototype, 'Update')
+  it("should awake and update all children", () => {
+    const spyNodeAwake = jest.spyOn(Node.prototype, "Awake")
+    const spyNodeUpdate = jest.spyOn(Node.prototype, "Update")
 
     expect(spyNodeAwake).not.toBeCalled()
     expect(spyNodeUpdate).not.toBeCalled()
