@@ -2,7 +2,6 @@ import "../mainPage/main-page.scss"
 import Typed from "typed.js"
 
 export default class MainPage {
-
   player1Nickname = ""
   player2Nickname = ""
   is2Players = false
@@ -10,20 +9,28 @@ export default class MainPage {
   player2Input: Nullable<HTMLInputElement> | undefined
   is2PlayersCheckbox: Nullable<HTMLInputElement> | undefined
   playersForm: Nullable<HTMLFormElement> | undefined
-  descriptions: string[] = ["- Brick walls -", "- Collect bonuses -", " - Explode your enemies -"]
+  descriptions: string[] = [
+    "- Brick walls -",
+    "- Collect bonuses -",
+    " - Explode your enemies -",
+  ]
   typedOptions = {
     strings: [...this.descriptions],
     typeSpeed: 40,
     loop: true,
     smartBackspace: true,
-    backSpeed: 40
+    backSpeed: 40,
   }
 
   private initVariables = (): void => {
     this.player1Input = document.querySelector("#player1") as HTMLInputElement
     this.player2Input = document.querySelector("#player2") as HTMLInputElement
-    this.is2PlayersCheckbox = document.querySelector("#numberOfPlayers") as HTMLInputElement
-    this.playersForm = document.querySelector(".player__form") as HTMLFormElement
+    this.is2PlayersCheckbox = document.querySelector(
+      "#numberOfPlayers"
+    ) as HTMLInputElement
+    this.playersForm = document.querySelector(
+      ".player__form"
+    ) as HTMLFormElement
   }
 
   private initEventListeners = (): void => {
@@ -47,7 +54,6 @@ export default class MainPage {
       // @ts-ignore
       this.player2Nickname = e.target.value
     })
-
   }
 
   private checkIs2Players = (): void => {
@@ -55,7 +61,9 @@ export default class MainPage {
       // eslint-disable-next-line @typescript-eslint/ban-ts-comment
       // @ts-ignore
       this.is2Players = e.target.checked
-      this.is2Players ? this.player2Input!.classList.add("player__name--show") : this.player2Input!.classList.remove("player__name--show")
+      this.is2Players
+        ? this.player2Input!.classList.add("player__name--show")
+        : this.player2Input!.classList.remove("player__name--show")
     })
   }
 
@@ -72,10 +80,9 @@ export default class MainPage {
     })
   }
 
-  render (): void{
+  render(): void {
     this.initVariables()
     this.initEventListeners()
     const typed = new Typed("#description", this.typedOptions)
   }
-
 }
