@@ -1,7 +1,7 @@
-import { Color } from "../color";
+import { Color } from "../color"
 
-describe('>>> Color', () => {
-  it('should instantiate with provided values', () => {
+describe(">>> Color", () => {
+  it("should instantiate with provided values", () => {
     const rgba = new Color(1, 2, 3, 0.1)
     expect(rgba.R).toEqual(1)
     expect(rgba.G).toEqual(2)
@@ -9,7 +9,7 @@ describe('>>> Color', () => {
     expect(rgba.A).toEqual(0.1)
   })
 
-  it('should throw an error if provided values are incorrect', () => {
+  it("should throw an error if provided values are incorrect", () => {
     expect(() => new Color(266, 2, 3, 0.1)).toThrowError(/Red/)
     expect(() => new Color(-1, 2, 3, 0.1)).toThrow(/Red/)
     expect(() => new Color(1.3, 2, 3, 0.1)).toThrow(/Red/)
@@ -26,25 +26,25 @@ describe('>>> Color', () => {
     expect(() => new Color(255, 255, 255, 1.2)).toThrowError(/Alpha/)
   })
 
-  it('should convert to string', () => {
+  it("should convert to string", () => {
     const rgba = new Color(1, 2, 3, 0.1)
-    expect(rgba.AsString()).toBe('rgba(1, 2, 3, 0.1)')
+    expect(rgba.AsString()).toBe("rgba(1, 2, 3, 0.1)")
   })
 
-  it('should instantiate from string', () => {
-    const rgba = Color.FromString('rgba(1, 2, 3, 0.1)')
+  it("should instantiate from string", () => {
+    const rgba = Color.FromString("rgba(1, 2, 3, 0.1)")
     expect(rgba.R).toEqual(1)
     expect(rgba.G).toEqual(2)
     expect(rgba.B).toEqual(3)
     expect(rgba.A).toEqual(0.1)
   })
 
-  it('should throw an error if cannot instantiate from string', () => {
-    expect(() => Color.FromString('')).toThrow()
-    expect(() => Color.FromString('?')).toThrow()
-    expect(() => Color.FromString('rgba()')).toThrow()
-    expect(() => Color.FromString('rgba(1)')).toThrow()
-    expect(() => Color.FromString('rgba(1,2)')).toThrow()
-    expect(() => Color.FromString('rgba(1,2,3)')).toThrow()
+  it("should throw an error if cannot instantiate from string", () => {
+    expect(() => Color.FromString("")).toThrow()
+    expect(() => Color.FromString("?")).toThrow()
+    expect(() => Color.FromString("rgba()")).toThrow()
+    expect(() => Color.FromString("rgba(1)")).toThrow()
+    expect(() => Color.FromString("rgba(1,2)")).toThrow()
+    expect(() => Color.FromString("rgba(1,2,3)")).toThrow()
   })
 })
