@@ -1,1 +1,1 @@
-export * from './color'
+export * from "./color"
