@@ -1,6 +1,6 @@
-import { Entity } from './entity'
-import { IAwake, IUpdate } from "../lifecycle";
+import { Entity } from "./entity"
+import { IAwake, IUpdate } from "../lifecycle"
 
 export interface IComponent extends IUpdate, IAwake {
-    Entity: Entity | null
+  Entity: Entity | null
 }
