@@ -1,2 +1,2 @@
-export * from './component.h'
-export * from './entity'
+export * from "./component.h"
+export * from "./entity"
