@@ -1,5 +1,5 @@
-export * from './ecs'
-export * from './vector2D'
-export * from './canvas'
-export * from './lifecycle'
-export * from './color'
+export * from "./ecs"
+export * from "./vector2D"
+export * from "./canvas"
+export * from "./lifecycle"
+export * from "./color"
