@@ -1,3 +1,3 @@
 export class Vector2D {
-    constructor(public x: number, public y: number) { }
+  constructor(public x: number, public y: number) {}
 }
